(SIMP-8459) Preserve ownership of rsync'd files
This patch addresses a recently-introduced bug that removed all
ownership, permissions, and SELinux contexts from rsync'd tftpboot image
files.  This resulted in tftpboot template files with default
(locked-down and root-only) permissions, which ths tftpboot service
(group `nobody`) was unable to access.

This patch reverts to the original rsync behavior, which is intended to
preserve file permissons, ownership, and SELinux contexts through to the
destination.

[SIMP-8459] #close